refactor(utils): Move R2 utilities to r2/ subdirectory
Phase 1 (Day 1): Low-risk migration
- Moved 3 R2 files to src/utils/r2/
- Updated 1 import in src/api-v3/jobs/scans.ts
- Tests passing (199/199 smoke tests)
- Biome linter clean

Files moved:
- r2-utils.ts
- r2-lifecycle.ts
- r2-hibernation.ts
